# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "What to do API",
        "description": "What to do API which provide all needed informations in order to create an application that provide a random activity for the evening",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "paths": {
        "/user": {
            "get": {
                "x-swagger-router-controller": "userController",
                "operationId": "list_all_user",
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in" : "header",
                        "required" : false,
                        "type" : "string"
                    }
                ],
                "description": "Get all users",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            },
            "post": {
                "x-swagger-router-controller": "userController",
                "operationId": "create_a_user",
                "tags": [
                    "User"
                ],
                "description": "Create a user",
                "parameters": [
                    {
                        "name": "CreatedAt",
                        "in": "formData",
                        "required": true,
                        "format": "date",
                        "type": "string"
                    },
                    {
                        "name": "Login",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "Password",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "SteamLogin",
                        "in": "formData",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "SteamPassword",
                        "in": "formData",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "TwithLogin",
                        "in": "formData",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "TwitchPassword",
                        "in": "formData",
                        "type": "string",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "User object not valid",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            }
        },
        "/user/{userId}": {
            "get": {
                "x-swagger-router-controller": "userController",
                "operationId": "read_a_user",
                "tags": [
                    "User"
                ],
                "description": "Get user informations from the Id",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            },
            "put": {
                "x-swagger-router-controller": "userController",
                "operationId": "update_a_user",
                "tags": [
                    "User"
                ],
                "description": "Update user information from Id",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "CreatedAt",
                        "in": "formData",
                        "required": true,
                        "format": "date",
                        "type": "string"
                    },
                    {
                        "name": "Login",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "Password",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "SteamLogin",
                        "in": "formData",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "SteamPassword",
                        "in": "formData",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "TwithLogin",
                        "in": "formData",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "TwitchPassword",
                        "in": "formData",
                        "type": "string",
                        "required": false
                    }
                ],
                "responses": {}
            },
            "delete": {
                "x-swagger-router-controller": "userController",
                "operationId": "delete_a_user",
                "tags": [
                    "User"
                ],
                "description": "Delete user from Id",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {}
            }
        },
        "/user/{userId}/type": {
            "get": {
                "x-swagger-router-controller": "userController",
                "operationId": "list_type_user",
                "tags": [
                    "User"
                ],
                "description": "Get user type informations from the Id",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            },
            "post": {
                "x-swagger-router-controller": "userController",
                "operationId": "add_type_user",
                "tags": [
                    "User"
                ],
                "description": "Create a user type",
                "parameters": [
                    {
                        "name": "UserId",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "TypeId",
                        "in": "formData",
                        "type": "integer",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/TypeUser"
                        }
                    },
                    "400": {
                        "description": "Type user object not valid",
                        "schema": {
                            "$ref": "#/definitions/TypeUser"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            }
        },
        "/history": {
            "get": {
                "x-swagger-router-controller": "historyController",
                "operationId": "list_all_history",
                "tags": [
                    "History"
                ],
                "description": "Get all history",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/History"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            },
            "post": {
                "x-swagger-router-controller": "historyController",
                "operationId": "create_a_history",
                "tags": [
                    "History"
                ],
                "description": "Create a history",
                "parameters": [
                    {
                        "name": "CreatedAt",
                        "in": "formData",
                        "required": true,
                        "format": "date",
                        "type": "string"
                    },
                    {
                        "name": "IdUser",
                        "in": "formData",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "name": "Type",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "Name",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/History"
                        }
                    },
                    "400": {
                        "description": "History object not valid",
                        "schema": {
                            "$ref": "#/definitions/History"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            }
        },
        "/history/{historyId}": {
            "get": {
                "x-swagger-router-controller": "historyController",
                "operationId": "read_a_history",
                "tags": [
                    "History"
                ],
                "description": "Get history informations from the Id",
                "parameters": [
                    {
                        "name": "historyId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/History"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            },
            "put": {
                "x-swagger-router-controller": "historyController",
                "operationId": "update_a_history",
                "tags": [
                    "History"
                ],
                "description": "Update history information from Id",
                "parameters": [
                    {
                        "name": "historyId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "CreatedAt",
                        "in": "formData",
                        "required": true,
                        "format": "date",
                        "type": "string"
                    },
                    {
                        "name": "IdUser",
                        "in": "formData",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "name": "Type",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "Name",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "x-swagger-router-controller": "historyController",
                "operationId": "delete_a_history",
                "tags": [
                    "History"
                ],
                "description": "Delete history from Id",
                "parameters": [
                    {
                        "name": "historyId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {}
            }
        },
        "/type": {
            "get": {
                "x-swagger-router-controller": "typeController",
                "operationId": "list_all_type",
                "tags": [
                    "Type"
                ],
                "description": "Get all type",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Type"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            },
            "post": {
                "x-swagger-router-controller": "typeController",
                "operationId": "create_a_history",
                "tags": [
                    "Type"
                ],
                "description": "Create a history",
                "parameters": [
                    {
                        "name": "Type",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Name",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Type"
                        }
                    },
                    "400": {
                        "description": "Type object not valid",
                        "schema": {
                            "$ref": "#/definitions/Type"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            }
        },
        "/type/{typeId}": {
            "get": {
                "x-swagger-router-controller": "typeController",
                "operationId": "read_a_type",
                "tags": [
                    "Type"
                ],
                "description": "Get type informations from the Id",
                "parameters": [
                    {
                        "name": "typeId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Type"
                        }
                    },
                    "default": {
                        "$ref": "#/definitions/responses/GlobalError"
                    }
                }
            },
            "put": {
                "x-swagger-router-controller": "typeController",
                "operationId": "update_a_type",
                "tags": [
                    "Type"
                ],
                "description": "Update type information from Id",
                "parameters": [
                    {
                        "name": "typeId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Type",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Name",
                        "in": "formData",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "x-swagger-router-controller": "typeController",
                "operationId": "delete_a_type",
                "tags": [
                    "Type"
                ],
                "description": "Delete type from Id",
                "parameters": [
                    {
                        "name": "typeId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "responses": {
            "GlobalError": {
                "description": "Execution error"
            }
        },
        "User": {
            "required": [
                "CreatedAt",
                "Login",
                "Password",
                "SteamLogin",
                "SteamPassword",
                "TwithLogin",
                "TwitchPassword"
            ],
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "Login": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                },
                "SteamLogin": {
                    "type": "string"
                },
                "SteamPassword": {
                    "type": "string"
                },
                "TwithLogin": {
                    "type": "string"
                },
                "TwitchPassword": {
                    "type": "string"
                }
            }
        },
        "History": {
            "required": [
                "CreatedAt",
                "IdUser",
                "Type",
                "Name"
            ],
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "IdUser": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
        "Type": {
            "required": [
                "Type",
                "Name"
            ],
            "properties": {
                "Type": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
        "TypeUser": {
            "required": [
                "userId",
                "typeId"
            ],
            "properties": {
                "userId": {
                    "type": "integer"
                },
                "typeId": {
                    "type": "integer"
                }
            }
        }
    }
}